refactor(150512): :truck: move managers para terceirizada
