+++ src/terceirizada/managers.py
